Minor refactor routine tests
